Introduce isl C++ bindings - Part 1 (smart pointer interface)
Over the last couple of months several authors of independent isl C++ bindings
worked together to jointly design an official set of isl C++ bindings which
combines their experience in developing isl C++ bindings. The new bindings are
based around an isl C++ smart pointer interface and remove the need for explicit
pointer managenent and instead use C++ language features to manage isl objects.

This commit introduces the smart-pointer part of the isl C++ bindings and
replaces the current IslPtr<T> classes, which served the very same purpose,
but had to be manually maintained.

An isl object has the following smart pointer interface:

    inline Set manage(__isl_take isl_set *ptr);

    class Set {
      friend inline Set manage(__isl_take isl_set *ptr);
      isl_set *ptr = nullptr;
      inline explicit Set(__isl_take isl_set *ptr);

    public:
      inline Set();
      inline Set(const Set &obj);
      inline Set &operator=(Set obj);
      inline ~Set();
      inline __isl_give isl_set *copy() const &;
      inline __isl_give isl_set *copy() && = delete;
      inline __isl_keep isl_set *get() const;
      inline __isl_give isl_set *release();
      inline bool isNull() const;
    }

An isl_set* s is converted to an isl::Set by calling 'S = isl::manage(s)'.

S.copy() returns a raw pointer refering to a copy of S.
S.get() extracts the raw pointer of the object managed by S.
S.release() extracts the raw pointer of the object managed by S and sets the
object in S to null.

S.isNull() checks if S manages a pointer or if the managed object is currently
null.

This commit also introduces a couple of polly-specific extensions that cover
features currently not handled by the official isl C++ bindings draft, but
which have been provided by IslPtr<T> and are consequently added to avoid code
churn. These extensions include:

	- operator bool() : Conversion from objects to bool
	- construction from nullptr_t
	- getCtx() method
	- take/keep/give methods, which match the currently used naming
	  convention of IslPtr<T> in Polly. They just forward to
	  (release/get/manage).
	- raw_ostream printers

We expect that these extensions are over time either removed or upstreamed to
the official isl bindings.

We also export a couple of classes that have not yet been exported in isl.

These bindings are not yet officially part of isl, but the draft is already very
stable. The smart pointer interface itself did not change since serveral months.
Adding this code to Polly is against our normal policy of only importing
official isl code. In this case however, we make an exception to showcase a
non-trivial use case of these bindings which should increase confidence in these
bindings and will help upstreaming them to isl.