Use isl C++ foreach implementation
This commit switches Polly over to the isl::Obj::foreach_* implementation,
which is part of the new isl bindings and follows the foreach pattern
established in Polly by Michael Kruse.

The original isl C function:

isl_stat isl_union_set_foreach_set(__isl_keep isl_union_set *uset,
        isl_stat (*fn)(__isl_take isl_set *set, void *user), void *user);

which required the user to define a static callback function to which all
interesting parameters are passed via a 'void *' user-pointer, is on the
C++ side available as a function that takes a std::function<>, which can
carry any additional arguments without the need for a user pointer:

stat UnionSet::foreach_set(std::function<stat(set)> &&fn) const;

The following code illustrates the use of the new C++ interface:

    auto Lambda = [=, &Result](isl::set Set) -> isl::stat {
      auto Shifted = shiftDimension(Set, Pos, Amount);
      Result = Result.add(Shifted);
      return isl::stat::ok;
    }

    UnionSet.foreach_set(Lambda);

Polly had some specialized foreach functions which did not require the lambdas
to return a status flag. We remove these functions in this commit to move
Polly completely over to the new isl interface. We may in the future discuss
if functors without return values can be supported easily.